\name{EigenSmallCountRounding-package}
\alias{EigenSmallCountRounding-package}
\alias{EigenSmallCountRounding}
\docType{package}
\title{
\packageTitle{EigenSmallCountRounding}
}
\description{
\packageDescription{EigenSmallCountRounding}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EigenSmallCountRounding}
\packageIndices{EigenSmallCountRounding}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EigenSmallCountRounding}

Maintainer: \packageMaintainer{EigenSmallCountRounding}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
